Calculate similarity in redux actions
